+++ app/src/main/java/com/donnKey/aesopPlayer/model/AudioBookManager.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2018-2019 Donn S. Terry
+ * Copyright (c) 2018-2020 Donn S. Terry
  * Copyright (c) 2015-2017 Marcin Simonides
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -81,7 +81,10 @@ public class AudioBookManager {
     public void setCurrentBook(String bookId) {
         AudioBook newBook = getById(bookId);
         if (newBook != currentBook) {
-            currentBook = getById(bookId);
+            if (currentBook != null) {
+                currentBook.leave();
+            }
+            currentBook = newBook;
             EventBus.getDefault().post(new CurrentBookChangedEvent(currentBook));
         }
     }