+++ app/src/main/java/com/donnKey/aesopPlayer/model/DemoSamplesInstaller.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2018-2019 Donn S. Terry
+ * Copyright (c) 2018-2020 Donn S. Terry
  * Copyright (c) 2015-2017 Marcin Simonides
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -26,7 +26,6 @@ package com.donnKey.aesopPlayer.model;
 
 import android.content.Context;
 import android.media.MediaScannerConnection;
-import android.util.Log;
 
 import androidx.annotation.MainThread;
 import androidx.annotation.WorkerThread;